+++ beacon-chain/core/blocks/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["block_operations.go"],
+    srcs = [
+        "block_operations.go",
+        "validity_conditions.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/blocks",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
@@ -11,16 +14,25 @@ go_library(
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/params:go_default_library",
         "//shared/slices:go_default_library",
+        "@com_github_ethereum_go_ethereum//common:go_default_library",
+        "@com_github_ethereum_go_ethereum//core/types:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["block_operations_test.go"],
+    srcs = [
+        "block_operations_test.go",
+        "validity_conditions_test.go",
+    ],
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/core/types:go_default_library",
+        "//beacon-chain/utils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/params:go_default_library",
+        "@com_github_ethereum_go_ethereum//common:go_default_library",
+        "@com_github_ethereum_go_ethereum//core/types:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
