--- beacon-chain/core/state/validity_conditions.go
@@ -1,181 +1,0 @@
-package state
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	"github.com/ethereum/go-ethereum/common"
-	gethTypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/types"
-	v "github.com/prysmaticlabs/prysm/beacon-chain/core/validators"
-	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
-	"github.com/prysmaticlabs/prysm/shared/bitutil"
-	"github.com/prysmaticlabs/prysm/shared/params"
-)
-
-// IsValidBlock ensures that the block is compliant with the block processing validity conditions.
-// Spec:
-//  For a beacon chain block, block, to be processed by a node, the following conditions must be met:
-//  The parent block with root block.parent_root has been processed and accepted.
-//  The node has processed its state up to slot, block.slot - 1.
-//  The Ethereum 1.0 block pointed to by the state.processed_pow_receipt_root has been processed and accepted.
-//  The node's local clock time is greater than or equal to state.genesis_time + block.slot * SLOT_DURATION.
-func IsValidBlock(
-	ctx context.Context,
-	state *types.BeaconState,
-	block *types.Block,
-	enablePOWChain bool,
-	HasBlock func(hash [32]byte) bool,
-	GetPOWBlock func(ctx context.Context, hash common.Hash) (*gethTypes.Block, error),
-	genesisTime time.Time) error {
-
-	// Pre-Processing Condition 1:
-	// Check that the parent Block has been processed and saved.
-	parentBlock := HasBlock(block.ParentHash())
-	if !parentBlock {
-		return fmt.Errorf("unprocessed parent block as it is not saved in the db: %#x", block.ParentHash())
-	}
-
-	// Pre-Processing Condition 2:
-	// The state is updated up to block.slot -1.
-
-	if state.Slot() != block.SlotNumber()-1 {
-		return fmt.Errorf(
-			"block slot is not valid %d as it is supposed to be %d", block.SlotNumber(), state.Slot()+1)
-	}
-
-	if enablePOWChain {
-		powBlock, err := GetPOWBlock(ctx, state.ProcessedPowReceiptRootHash32())
-		if err != nil {
-			return fmt.Errorf("unable to retrieve POW chain reference block %v", err)
-		}
-
-		// Pre-Processing Condition 3:
-		// The block pointed to by the state in state.processed_pow_receipt_root has
-		// been processed in the ETH 1.0 chain.
-		if powBlock == nil {
-			return fmt.Errorf("proof-of-Work chain reference in state does not exist %#x", state.ProcessedPowReceiptRootHash32())
-		}
-	}
-
-	// Pre-Processing Condition 4:
-	// The node's local time is greater than or equal to
-	// state.genesis_time + block.slot * SLOT_DURATION.
-	if !block.IsSlotValid(genesisTime) {
-		return fmt.Errorf("slot of block is too high: %d", block.SlotNumber())
-	}
-
-	return nil
-}
-
-// doesParentProposerExist checks that the proposer from the parent slot is included in the first
-// aggregated attestation object
-func doesParentProposerExist(block *types.Block, beaconState *types.BeaconState, parentSlot uint64) error {
-	_, parentProposerIndex, err := v.ProposerShardAndIndex(
-		beaconState.ShardAndCommitteesForSlots(),
-		beaconState.LastStateRecalculationSlot(),
-		parentSlot,
-	)
-	if err != nil {
-		return err
-	}
-
-	// Verifies the attester bitfield to check if the proposer index is in the first included one.
-	if isBitSet, err := bitutil.CheckBit(block.Attestations()[0].AttesterBitfield, int(parentProposerIndex)); !isBitSet {
-		return fmt.Errorf("could not locate proposer in the first attestation of AttestionRecord: %v", err)
-	}
-	return nil
-}
-
-// isBlockAttestationValid verifies a block's attestations pass validity conditions.
-// TODO(#781): Refactor with the new spec attestation checking conditions.
-func isBlockAttestationValid(
-	block *types.Block,
-	attestation *pb.AggregatedAttestation,
-	beaconState *types.BeaconState,
-	parentSlot uint64,
-	isInChain func(blockHash [32]byte) bool,
-) error {
-	// Validate attestation's slot number has is within range of incoming block number.
-	if err := isAttestationSlotNumberValid(attestation.Slot, parentSlot); err != nil {
-		return fmt.Errorf("invalid attestation slot %d: %v", attestation.Slot, err)
-	}
-
-	if attestation.JustifiedSlot > beaconState.LastJustifiedSlot() {
-		return fmt.Errorf(
-			"attestation's justified slot has to be <= the state's last justified slot: found: %d. want <=: %d",
-			attestation.JustifiedSlot,
-			beaconState.LastJustifiedSlot(),
-		)
-	}
-
-	hash := [32]byte{}
-	copy(hash[:], attestation.JustifiedBlockHash)
-	if !isInChain(hash) {
-		return fmt.Errorf(
-			"the attestation's justifed block hash not found in current chain: justified block hash: 0x%x",
-			attestation.JustifiedBlockHash,
-		)
-	}
-
-	// Get all the block hashes up to cycle length.
-	parentHashes, err := beaconState.SignedParentHashes(block, attestation)
-	if err != nil {
-		return fmt.Errorf("unable to get signed parent hashes: %v", err)
-	}
-
-	shardCommittees, err := v.GetShardAndCommitteesForSlot(
-		beaconState.ShardAndCommitteesForSlots(),
-		beaconState.LastStateRecalculationSlot(),
-		attestation.Slot,
-	)
-	attesterIndices, err := v.AttesterIndices(shardCommittees, attestation)
-	if err != nil {
-		return fmt.Errorf("unable to get validator committee: %v", err)
-	}
-
-	// Verify attester bitfields matches crystallized state's prev computed bitfield.
-	if !v.AreAttesterBitfieldsValid(attestation, attesterIndices) {
-		return fmt.Errorf("unable to match attester bitfield with shard and committee bitfield")
-	}
-
-	// TODO(#258): Add coverage for determining fork version for an attestation.
-	forkData := beaconState.ForkData()
-	forkVersion := forkData.PostForkVersion
-	if attestation.Slot < forkData.ForkSlot {
-		forkVersion = forkData.PreForkVersion
-	}
-
-	// TODO(#258): Generate validators aggregated pub key.
-	attestationMsg := types.AttestationMsg(
-		parentHashes,
-		attestation.ShardBlockHash,
-		attestation.Slot,
-		attestation.Shard,
-		attestation.JustifiedSlot,
-		forkVersion,
-	)
-	_ = attestationMsg
-
-	// TODO(#258): Verify msgHash against aggregated pub key and aggregated signature.
-	return nil
-}
-
-func isAttestationSlotNumberValid(attestationSlot uint64, parentSlot uint64) error {
-	if parentSlot != 0 && attestationSlot > parentSlot {
-		return fmt.Errorf(
-			"attestation slot number higher than parent block's slot number: found: %d, needed < %d",
-			attestationSlot,
-			parentSlot,
-		)
-	}
-	if parentSlot >= params.BeaconConfig().CycleLength-1 && attestationSlot < parentSlot-params.BeaconConfig().CycleLength+1 {
-		return fmt.Errorf(
-			"attestation slot number lower than parent block's slot number by one CycleLength: found: %d, needed > %d",
-			attestationSlot,
-			parentSlot-params.BeaconConfig().CycleLength+1,
-		)
-	}
-	return nil
-}
